+++ src/pages/Gear/GearItemPage.tsx
@@ -3,11 +3,10 @@ import { Notes } from 'components/Notes';
 import { useSetPageTitle } from 'hooks';
 import { useParams } from 'react-router-dom';
 import { useGetGearItemQuery } from 'redux/api';
-import styled from 'styled-components';
 
 import { GearInfoPanel } from './GearInfoPanel';
-import { GearPicturePicker } from './GearPicturePicker';
 import { GearRentalsHistory } from './GearRentalsHistory';
+import { GearPicture } from './GearPicture';
 
 export function GearItemPage() {
   const { gearId } = useParams<{ gearId: string }>();
@@ -21,11 +20,7 @@ export function GearItemPage() {
     <div className="row">
       <div className="col-12 col-md-5 p-2">
         <GearInfoPanel gearItem={gearItem} refreshGear={refreshGear} />
-        {gearItem.picture ? (
-          <GearPicture src={gearItem.picture} alt="Gear item" />
-        ) : (
-          <GearPicturePicker item={gearItem} refreshGear={refreshGear} />
-        )}
+        <GearPicture gearItem={gearItem} refreshGear={refreshGear} />
         <Notes notes={gearItem.notes} onAdd={(note) => addNote(gearId, note).then(refreshGear)} />
       </div>
       <div className="col-12 col-md-7 p-2">
@@ -34,7 +29,3 @@ export function GearItemPage() {
     </div>
   );
 }
-
-const GearPicture = styled.img`
-  width: 100%;
-`;
